refactor: sep style
