+++ src/common/GoogleApp.tsx
@@ -1,329 +1,308 @@
-import React from "react";
-import styled from "styled-components";
+import React from 'react';
+import styled from 'styled-components';
+
+function GoogleApp() {
+	return (
+		<a
+			href="https://play.google.com/store/apps/details?id=com.stockeeper.stockeeperapp"
+			target="_blank"
+			rel="noreferrer"
+		>
+			<Box>
+				<svg
+					preserveAspectRatio="xMidYMid meet"
+					data-bbox="0.5 -115.6 443 177.1"
+					viewBox="0 0 208 62"
+					xmlns="http://www.w3.org/2000/svg"
+					data-type="ugc"
+					role="presentation"
+					aria-hidden="true"
+					aria-labelledby="svgcid-i4ic11-ni6gtn"
+				>
+					<title id="svgcid-i4ic11-ni6gtn"></title>
+					<g>
+						<path d="M16.1 49.9c-.4-.5-.7-1.2-.7-2.2v.2c0 1 .3 1.7.7 2.2l.1-.1-.1-.1z" opacity=".122"></path>
+						<path
+							d="M.5 51.5v-41c0-.7.1-1.3.2-2 .1-.6.3-1.2.6-1.8s.6-1.2.9-1.7c.4-.6.8-1.1 1.2-1.6.5-.5 1-.9 1.5-1.2.5-.4 1.1-.7 1.7-.9C7.3 1 7.9.8 8.5.7s1.3-.2 2-.2h187c.7 0 1.3.1 2 .2.6.1 1.3.3 1.9.6.6.3 1.2.6 1.7.9.5.4 1.1.8 1.5 1.2.5.5.9 1 1.2 1.5.4.5.7 1.1.9 1.7.3.6.4 1.2.6 1.9.1.6.2 1.3.2 2v41c0 .7-.1 1.3-.2 2-.1.6-.3 1.3-.6 1.9-.3.6-.6 1.2-.9 1.7s-.8 1.1-1.2 1.5c-.5.5-1 .9-1.5 1.2-.5.4-1.1.7-1.7.9-.6.3-1.2.4-1.9.6s-1.3.2-2 .2h-187c-.7 0-1.3-.1-2-.2s-1.2-.3-1.8-.6c-.6-.3-1.2-.6-1.7-.9-.5-.4-1.1-.8-1.5-1.2-.5-.5-.9-1-1.2-1.5s-.7-1.1-.9-1.7c-.3-.6-.4-1.2-.6-1.9-.2-.7-.3-1.3-.3-2z"
+							fill="#ffffff"
+						></path>
+						<path
+							d="M.5 52V10c0-.6.1-1.2.2-1.9.1-.6.3-1.2.5-1.7.2-.6.5-1.1.9-1.6s.7-1 1.2-1.4.9-.8 1.4-1.2 1.1-.6 1.6-.9C6.9 1 7.5.8 8.1.7S9.4.5 10 .5h188c.6 0 1.2.1 1.9.2s1.2.3 1.8.5c.6.2 1.1.5 1.6.9s1 .7 1.4 1.2c.4.4.8.9 1.2 1.4s.6 1.1.9 1.6c.2.6.4 1.2.5 1.8.1.6.2 1.2.2 1.9v42c0 .6-.1 1.2-.2 1.9s-.3 1.2-.5 1.8c-.2.6-.5 1.1-.9 1.6-.3.5-.7 1-1.2 1.4-.4.4-.9.8-1.4 1.2-.5.3-1.1.6-1.6.9-.6.2-1.2.4-1.8.5-.6.1-1.2.2-1.9.2H10c-.6 0-1.2-.1-1.9-.2-.6-.1-1.2-.3-1.8-.5s-1.1-.5-1.6-.9c-.5-.3-1-.7-1.4-1.2-.4-.4-.8-.9-1.2-1.4s-.6-1.1-.9-1.6c-.2-.6-.4-1.2-.5-1.8-.1-.7-.2-1.3-.2-1.9z"
+							fill="none"
+							stroke="#a5a5a5"
+						></path>
+						<path d="M164.2 39.8h2.9V20.6h-2.9v19.2zM190 27.5l-3.3 8.4h-.1l-3.4-8.4h-3.1l5.1 11.7-2.9 6.5h3l7.9-18.2H190zm-16.3 10.2c-.9 0-2.3-.5-2.3-1.6 0-1.5 1.6-2.1 3-2.1 1.3 0 1.9.3 2.6.6-.1 1.8-1.6 3.1-3.3 3.1zm.4-10.6c-2.1 0-4.2.9-5.1 2.9l2.6 1.1c.5-1.1 1.6-1.4 2.6-1.4 1.5 0 3 .9 3 2.5v.2c-.5-.3-1.6-.7-3-.7-2.8 0-5.6 1.5-5.6 4.3 0 2.6 2.3 4.2 4.8 4.2 1.9 0 3-.9 3.7-1.9h.1v1.5h2.8v-7.4c-.1-3.4-2.6-5.3-5.9-5.3zm-17.8 2.8h-4.1v-6.6h4.1c2.1 0 3.4 1.8 3.4 3.3 0 1.5-1.2 3.3-3.4 3.3zm-.1-9.3h-6.9v19.3h2.9v-7.3h4c3.2 0 6.3-2.3 6.3-6 .1-3.7-3.1-6-6.3-6zm-37.5 17.1c-2 0-3.6-1.7-3.6-3.9 0-2.3 1.7-4 3.6-4 2 0 3.5 1.7 3.5 4 0 2.2-1.6 3.9-3.5 3.9zm3.3-9.1c-.7-.8-2-1.5-3.5-1.5-3.3 0-6.3 2.9-6.3 6.6 0 3.7 3 6.5 6.3 6.5 1.6 0 2.8-.7 3.4-1.5h.1v.9c0 2.5-1.3 3.8-3.5 3.8-1.8 0-2.9-1.3-3.3-2.3l-2.5 1c.7 1.7 2.6 3.9 5.8 3.9 3.4 0 6.2-2 6.2-6.8V27.5H122v1.1zm4.7 11.2h2.9V20.6h-2.9v19.2zm7.1-6.3c-.1-2.5 2-3.8 3.4-3.8 1.1 0 2.1.6 2.4 1.4l-5.8 2.4zm9-2.2c-.5-1.5-2.2-4.2-5.6-4.2-3.4 0-6.2 2.7-6.2 6.6 0 3.7 2.8 6.6 6.5 6.6 3 0 4.7-1.8 5.5-2.9l-2.2-1.5c-.7 1.1-1.8 1.8-3.2 1.8-1.5 0-2.5-.7-3.2-2l8.8-3.6-.4-.8zM73 29.1v2.8h6.7c-.2 1.6-.7 2.7-1.5 3.5-1 1-2.5 2-5.1 2-4.1 0-7.3-3.3-7.3-7.4 0-4.1 3.2-7.4 7.3-7.4 2.2 0 3.8.9 5 2l2-2c-1.7-1.6-3.9-2.8-7-2.8-5.6 0-10.3 4.6-10.3 10.2s4.7 10.2 10.3 10.2c3 0 5.3-1 7.1-2.9 1.8-1.8 2.4-4.4 2.4-6.5 0-.6 0-1.2-.1-1.7H73zm17 8.6c-2 0-3.7-1.6-3.7-4s1.7-4 3.7-4 3.7 1.6 3.7 4c0 2.3-1.7 4-3.7 4zm0-10.6c-3.6 0-6.6 2.8-6.6 6.6s3 6.6 6.6 6.6c3.6 0 6.6-2.8 6.6-6.6s-3-6.6-6.6-6.6zm14.4 10.6c-2 0-3.7-1.6-3.7-4s1.7-4 3.7-4 3.7 1.6 3.7 4c0 2.3-1.7 4-3.7 4zm0-10.6c-3.6 0-6.6 2.8-6.6 6.6s3 6.6 6.6 6.6c3.6 0 6.6-2.8 6.6-6.6-.1-3.8-3-6.6-6.6-6.6z"></path>
+						<g transform="translate(25.366 20.986)">
+							<linearGradient
+								gradientTransform="matrix(19.9196 0 0 -38.628 6492.775 7042.514)"
+								y2="181.09"
+								x2="-324.559"
+								y1="180.421"
+								x1="-323.261"
+								gradientUnits="userSpaceOnUse"
+								id="svgcid-a3x0nd5iltcx"
+							>
+								<stop offset="0" stopColor="#00a0ff"></stop>
+								<stop offset=".007" stopColor="#00a1ff"></stop>
+								<stop offset=".26" stopColor="#00beff"></stop>
+								<stop offset=".512" stopColor="#00d2ff"></stop>
+								<stop offset=".76" stopColor="#00dfff"></stop>
+								<stop offset="1" stopColor="#00e3ff"></stop>
+							</linearGradient>
+							<path
+								d="M-9.9-9.3c-.5.6-.8 1.4-.7 2.2V27c0 .8.2 1.6.7 2.2l.1.1L9.3 10.1v-.4L-9.8-9.4l-.1.1z"
+								fill="url(#svgcid-a3x0nd5iltcx)"
+							></path>
+							<linearGradient
+								gradientTransform="matrix(15.661 0 0 -13.181 5054.71 2158.078)"
+								y2="162.976"
+								x2="-323.468"
+								y1="162.976"
+								x1="-321.087"
+								gradientUnits="userSpaceOnUse"
+								id="svgcid--g2g3vz-5nrlpi"
+							>
+								<stop offset="0" stopColor="#ffe000"></stop>
+								<stop offset=".409" stopColor="#ffbd00"></stop>
+								<stop offset=".775" stopColor="#ffa500"></stop>
+								<stop offset="1" stopColor="#ff9c00"></stop>
+							</linearGradient>
+							<path
+								d="m15.7 16.5-6.4-6.4v-.4l6.4-6.4.1.1 7.5 4.3c2.2 1.2 2.2 3.2 0 4.4l-7.5 4.3-.1.1z"
+								fill="url(#svgcid--g2g3vz-5nrlpi)"
+							></path>
+							<linearGradient
+								gradientTransform="matrix(25.708 0 0 -19.8124 8475.259 3440.851)"
+								y2="167.706"
+								x2="-325.849"
+								y1="165.936"
+								x1="-324.486"
+								gradientUnits="userSpaceOnUse"
+								id="svgcid-11ngs6-9ko6u8"
+							>
+								<stop offset="0" stopColor="#ff3a44"></stop>
+								<stop offset="1" stopColor="#c31162"></stop>
+							</linearGradient>
+							<path d="M15.8 16.4 9.3 9.9-9.9 29.1c.7.8 1.9.8 3.2.1l22.5-12.8" fill="url(#svgcid-11ngs6-9ko6u8)"></path>
+							<linearGradient
+								gradientTransform="matrix(25.707 0 0 -19.8124 8474.918 3421.039)"
+								y2="164.979"
+								x2="-335.604"
+								y1="164.189"
+								x1="-336.213"
+								gradientUnits="userSpaceOnUse"
+								id="svgcid-4cpanjacuzsy"
+							>
+								<stop offset="0" stopColor="#32a071"></stop>
+								<stop offset=".068" stopColor="#2da771"></stop>
+								<stop offset=".476" stopColor="#15cf74"></stop>
+								<stop offset=".801" stopColor="#06e775"></stop>
+								<stop offset="1" stopColor="#00f076"></stop>
+							</linearGradient>
+							<path d="M15.8 3.4-6.7-9.4c-1.3-.8-2.5-.7-3.2.1L9.3 9.9l6.5-6.5z" fill="url(#svgcid-4cpanjacuzsy)"></path>
+							<path
+								d="m15.7 16.279-22.4 12.7c-.9.7-2.2.7-3.1 0l-.1.1.1.1c.9.7 2.2.7 3.1 0l22.5-12.8-.1-.1z"
+								opacity=".2"
+							></path>
+							<path d="M-9.9 28.879c-.5-.6-.8-1.4-.7-2.2v.2c0 .8.2 1.6.7 2.2l.1-.1-.1-.1z" opacity=".12"></path>
+							<path d="m23.4 11.9-7.7 4.4.1.1 7.5-4.3c.9-.4 1.5-1.2 1.6-2.2-.1.9-.7 1.6-1.5 2z" opacity=".12"></path>
+							<path
+								d="m-6.7-9.2 30 17.1c.8.4 1.4 1.1 1.6 2-.1-1-.7-1.8-1.6-2.2l-30-17.1c-2.2-1.2-3.9-.2-3.9 2.3v.2c0-2.5 1.8-3.5 3.9-2.3z"
+								opacity=".25"
+								fill="#ffffff"
+							></path>
+						</g>
+						<path
+							d="M.5-64.1v-41c0-.7.1-1.3.2-2 .1-.6.3-1.3.6-1.9s.6-1.2.9-1.7c.4-.5.8-1.1 1.2-1.5.5-.5 1-.9 1.5-1.2.5-.4 1.1-.7 1.7-.9s1.2-.4 1.9-.6 1.3-.2 2-.2h187c.7 0 1.3.1 2 .2.6.1 1.3.3 1.9.6.6.3 1.2.6 1.7.9.5.4 1.1.8 1.5 1.2.5.5.9 1 1.2 1.5.4.5.7 1.1.9 1.7.3.6.4 1.2.6 1.9.1.6.2 1.3.2 2v41c0 .7-.1 1.3-.2 2-.1.6-.3 1.3-.6 1.9-.3.6-.6 1.2-.9 1.7s-.8 1.1-1.2 1.5c-.5.5-1 .9-1.5 1.2-.5.4-1.1.7-1.7.9-.6.3-1.2.4-1.9.6s-1.3.2-2 .2h-187c-.7 0-1.3-.1-2-.2s-1.3-.3-1.9-.6c-.6-.3-1.2-.6-1.7-.9-.5-.4-1.1-.8-1.5-1.2-.5-.5-.9-1-1.2-1.5s-.7-1.1-.9-1.7c-.3-.6-.4-1.2-.6-1.9-.1-.7-.2-1.3-.2-2z"
+							fill="#ffffff"
+						></path>
+						<path
+							d="M.5-63.6v-42c0-.6.1-1.2.2-1.9s.3-1.2.5-1.8.5-1.1.9-1.6.7-1 1.2-1.4.9-.8 1.4-1.2c.5-.3 1.1-.6 1.6-.9.6-.2 1.2-.4 1.8-.5.6-.1 1.2-.2 1.9-.2h188c.6 0 1.2.1 1.9.2.6.1 1.2.3 1.8.5.6.2 1.1.5 1.6.9.5.3 1 .7 1.4 1.2.4.4.8.9 1.2 1.4s.6 1.1.9 1.6.4 1.2.5 1.8.2 1.2.2 1.9v42c0 .6-.1 1.2-.2 1.9s-.3 1.2-.5 1.8c-.2.6-.5 1.1-.9 1.6-.3.5-.7 1-1.2 1.4-.4.4-.9.8-1.4 1.2-.5.3-1.1.6-1.6.9-.6.2-1.2.4-1.8.5-.6.1-1.2.2-1.9.2H10c-.6 0-1.2-.1-1.9-.2-.6-.1-1.2-.3-1.8-.5s-1.1-.5-1.6-.9c-.5-.3-1-.7-1.4-1.2-.4-.4-.8-.9-1.2-1.4s-.6-1.1-.9-1.6c-.2-.6-.4-1.2-.5-1.8-.1-.6-.2-1.2-.2-1.9z"
+							fill="none"
+							stroke="#a5a5a5"
+						></path>
+						<path d="M73.1-99.7c0 1.3-.4 2.3-1.1 3.1-.9.9-2 1.4-3.4 1.4-1.3 0-2.5-.5-3.4-1.4-.9-.9-1.4-2.1-1.4-3.4 0-1.4.5-2.5 1.4-3.4.9-.9 2.1-1.4 3.4-1.4.7 0 1.3.1 1.9.4.6.3 1.1.6 1.4 1l-.8.8c-.6-.7-1.5-1.1-2.5-1.1s-1.8.3-2.5 1-1.1 1.6-1.1 2.7.4 2 1.1 2.7c.7.7 1.5 1 2.5 1s1.9-.3 2.6-1c.4-.4.7-1.1.8-1.9h-3.4v-1.1H73c0 .1.1.3.1.6z"></path>
+						<path
+							d="M73.1-99.7c0 1.3-.4 2.3-1.1 3.1-.9.9-2 1.4-3.4 1.4-1.3 0-2.5-.5-3.4-1.4-.9-.9-1.4-2.1-1.4-3.4 0-1.4.5-2.5 1.4-3.4.9-.9 2.1-1.4 3.4-1.4.7 0 1.3.1 1.9.4.6.3 1.1.6 1.4 1l-.8.8c-.6-.7-1.5-1.1-2.5-1.1s-1.8.3-2.5 1-1.1 1.6-1.1 2.7.4 2 1.1 2.7c.7.7 1.5 1 2.5 1s1.9-.3 2.6-1c.4-.4.7-1.1.8-1.9h-3.4v-1.1H73c0 .1.1.3.1.6z"
+							fill="none"
+							stroke="#ffffff"
+							strokeWidth=".2"
+							strokeMiterlimit="10"
+						></path>
+						<path d="M80.2-103.6H76v2.9h3.7v1.2H76v2.9h4.2v1.1h-5.4v-9.2h5.4v1.1z"></path>
+						<path
+							fill="none"
+							stroke="#ffffff"
+							strokeWidth=".2"
+							strokeMiterlimit="10"
+							d="M80.2-103.6H76v2.9h3.7v1.2H76v2.9h4.2v1.1h-5.4v-9.2h5.4v1.1z"
+						></path>
+						<path d="M85.2-95.5H84v-8.1h-2.6v-1.1h6.4v1.1h-2.6v8.1z"></path>
+						<path
+							fill="none"
+							stroke="#ffffff"
+							strokeWidth=".2"
+							strokeMiterlimit="10"
+							d="M85.2-95.5H84v-8.1h-2.6v-1.1h6.4v1.1h-2.6v8.1z"
+						></path>
+						<path d="M93.5-104.7v9.2h-1.2v-9.2h1.2z"></path>
+						<path
+							fill="none"
+							stroke="#ffffff"
+							strokeWidth=".2"
+							strokeMiterlimit="10"
+							d="M93.5-104.7v9.2h-1.2v-9.2h1.2z"
+						></path>
+						<path d="M98.8-95.5h-1.2v-8.1H95v-1.1h6.4v1.1h-2.6v8.1z"></path>
+						<path
+							fill="none"
+							stroke="#ffffff"
+							strokeWidth=".2"
+							strokeMiterlimit="10"
+							d="M98.8-95.5h-1.2v-8.1H95v-1.1h6.4v1.1h-2.6v8.1z"
+						></path>
+						<path d="M107.5-97.4c.7.7 1.5 1 2.5 1s1.8-.3 2.5-1 1-1.6 1-2.7-.3-2-1-2.7c-.7-.7-1.5-1-2.5-1s-1.8.3-2.5 1-1 1.6-1 2.7.3 2 1 2.7zm5.9.7c-.9.9-2 1.4-3.4 1.4s-2.5-.5-3.4-1.4c-.9-.9-1.4-2.1-1.4-3.4 0-1.4.5-2.5 1.4-3.4.9-.9 2-1.4 3.4-1.4 1.3 0 2.5.5 3.4 1.4.9.9 1.4 2.1 1.4 3.4 0 1.3-.5 2.5-1.4 3.4z"></path>
+						<path
+							d="M107.5-97.4c.7.7 1.5 1 2.5 1s1.8-.3 2.5-1 1-1.6 1-2.7-.3-2-1-2.7c-.7-.7-1.5-1-2.5-1s-1.8.3-2.5 1-1 1.6-1 2.7.3 2 1 2.7zm5.9.7c-.9.9-2 1.4-3.4 1.4s-2.5-.5-3.4-1.4c-.9-.9-1.4-2.1-1.4-3.4 0-1.4.5-2.5 1.4-3.4.9-.9 2-1.4 3.4-1.4 1.3 0 2.5.5 3.4 1.4.9.9 1.4 2.1 1.4 3.4 0 1.3-.5 2.5-1.4 3.4z"
+							fill="none"
+							stroke="#ffffff"
+							strokeWidth=".2"
+							strokeMiterlimit="10"
+						></path>
+						<path d="M116.4-95.5v-9.2h1.5l4.5 7.2v-7.2h1.2v9.2h-1.3l-4.7-7.5v7.5h-1.2z"></path>
+						<path
+							fill="none"
+							stroke="#ffffff"
+							strokeWidth=".2"
+							strokeMiterlimit="10"
+							d="M116.4-95.5v-9.2h1.5l4.5 7.2v-7.2h1.2v9.2h-1.3l-4.7-7.5v7.5h-1.2z"
+						></path>
+						<path d="M164.8-69.3h2.9v-19.3h-2.9v19.3zm25.9-12.3-3.3 8.4h-.1l-3.4-8.4h-3.1l5.1 11.7-2.9 6.5h3l7.9-18.2h-3.2zm-16.3 10.1c-.9 0-2.3-.5-2.3-1.6 0-1.5 1.6-2.1 3-2.1 1.3 0 1.9.3 2.6.6-.2 1.9-1.7 3.1-3.3 3.1zm.3-10.5c-2.1 0-4.2.9-5.1 2.9l2.6 1.1c.5-1.1 1.6-1.4 2.6-1.4 1.5 0 3 .9 3 2.5v.2c-.5-.3-1.6-.7-3-.7-2.8 0-5.6 1.5-5.6 4.3 0 2.6 2.3 4.2 4.8 4.2 1.9 0 3-.9 3.7-1.9h.1v1.5h2.8v-7.4c-.1-3.4-2.6-5.3-5.9-5.3zm-17.8 2.7h-4.1v-6.6h4.1c2.1 0 3.4 1.8 3.4 3.3 0 1.5-1.2 3.3-3.4 3.3zm0-9.2H150v19.3h2.9v-7.3h4c3.2 0 6.3-2.3 6.3-6s-3.2-6-6.3-6zm-37.6 17c-2 0-3.6-1.7-3.6-3.9 0-2.3 1.7-4 3.6-4 2 0 3.5 1.7 3.5 4s-1.5 3.9-3.5 3.9zm3.3-9c-.7-.8-2-1.5-3.5-1.5-3.3 0-6.3 2.9-6.3 6.6 0 3.7 3 6.5 6.3 6.5 1.6 0 2.8-.7 3.4-1.5h.1v.9c0 2.5-1.3 3.8-3.5 3.8-1.8 0-2.9-1.3-3.3-2.3l-2.5 1c.7 1.7 2.6 3.9 5.8 3.9 3.4 0 6.2-2 6.2-6.8v-11.8h-2.7v1.2zm4.7 11.2h2.9v-19.3h-2.9v19.3zm7.1-6.3c-.1-2.5 2-3.8 3.4-3.8 1.1 0 2.1.6 2.4 1.4l-5.8 2.4zm9-2.2c-.5-1.5-2.2-4.2-5.6-4.2-3.4 0-6.2 2.7-6.2 6.6 0 3.7 2.8 6.6 6.5 6.6 3 0 4.7-1.8 5.5-2.9l-2.2-1.5c-.7 1.1-1.8 1.8-3.2 1.8-1.5 0-2.5-.7-3.2-2l8.8-3.6-.4-.8zM73.6-80v2.8h6.7c-.2 1.6-.7 2.7-1.5 3.5-1 1-2.5 2-5.1 2-4.1 0-7.3-3.3-7.3-7.4 0-4.1 3.2-7.4 7.3-7.4 2.2 0 3.8.9 5 2l2-2c-1.7-1.6-3.9-2.8-7-2.8-5.6 0-10.3 4.6-10.3 10.2S68-68.9 73.6-68.9c3 0 5.3-1 7.1-2.9 1.8-1.8 2.4-4.4 2.4-6.5 0-.6 0-1.2-.1-1.7h-9.4zm17 8.5c-2 0-3.7-1.6-3.7-4s1.7-4 3.7-4 3.7 1.6 3.7 4-1.7 4-3.7 4zm0-10.5C87-82 84-79.2 84-75.4s3 6.6 6.6 6.6c3.6 0 6.6-2.8 6.6-6.6S94.3-82 90.6-82zM105-71.5c-2 0-3.7-1.6-3.7-4s1.7-4 3.7-4 3.7 1.6 3.7 4-1.7 4-3.7 4zm0-10.5c-3.6 0-6.6 2.8-6.6 6.6s3 6.6 6.6 6.6c3.6 0 6.6-2.8 6.6-6.6s-3-6.6-6.6-6.6z"></path>
+						<linearGradient
+							gradientTransform="matrix(1 0 0 -1 -4.607 -175.041)"
+							y2="-191.279"
+							x2="121.728"
+							y1="-217.136"
+							x1="147.586"
+							gradientUnits="userSpaceOnUse"
+							id="svgcid--p8co1pbjv1nx"
+						>
+							<stop offset="0" stopColor="#00a0ff"></stop>
+							<stop offset=".007" stopColor="#00a1ff"></stop>
+							<stop offset=".26" stopColor="#00beff"></stop>
+							<stop offset=".512" stopColor="#00d2ff"></stop>
+							<stop offset=".76" stopColor="#00dfff"></stop>
+							<stop offset="1" stopColor="#00e3ff"></stop>
+						</linearGradient>
+						<path
+							d="M16.1-103.9c-.4.5-.7 1.2-.7 2.2v34.1c0 1 .3 1.7.7 2.2l.1.1 19.1-19.1v-.4L16.2-104l-.1.1z"
+							fill="url(#svgcid--p8co1pbjv1nx)"
+						></path>
+						<linearGradient
+							gradientTransform="matrix(1 0 0 -1 2.38 -170.576)"
+							y2="-85.885"
+							x2="12.467"
+							y1="-85.885"
+							x1="49.748"
+							gradientUnits="userSpaceOnUse"
+							id="svgcid--qj0fwh-s7c4e8"
+						>
+							<stop offset="0" stopColor="#ffe000"></stop>
+							<stop offset=".409" stopColor="#ffbd00"></stop>
+							<stop offset=".775" stopColor="#ffa500"></stop>
+							<stop offset="1" stopColor="#ff9c00"></stop>
+						</linearGradient>
+						<path
+							d="m41.6-78.1-6.4-6.4v-.4l6.4-6.4.1.1 7.5 4.3c2.2 1.2 2.2 3.2 0 4.4l-7.5 4.3-.1.1z"
+							fill="url(#svgcid--qj0fwh-s7c4e8)"
+						></path>
+						<linearGradient
+							gradientTransform="matrix(1 0 0 -1 -4.357 -175.217)"
+							y2="-201.946"
+							x2="-90.988"
+							y1="-217.604"
+							x1="-106.646"
+							gradientUnits="userSpaceOnUse"
+							id="svgcid-98kv97-7egd8e"
+						>
+							<stop offset="0" stopColor="#32a071"></stop>
+							<stop offset=".069" stopColor="#2da771"></stop>
+							<stop offset=".476" stopColor="#15cf74"></stop>
+							<stop offset=".801" stopColor="#06e775"></stop>
+							<stop offset="1" stopColor="#00f076"></stop>
+						</linearGradient>
+						<path
+							d="M41.8-91.2 19.3-104c-1.3-.8-2.5-.7-3.2.1l19.2 19.2 6.5-6.5z"
+							fill="url(#svgcid-98kv97-7egd8e)"
+						></path>
+						<path
+							d="M41.6-78.3 19.3-65.6c-1.3.7-2.4.7-3.1 0l-.1.1.1.1c.7.6 1.8.7 3.1 0l22.5-12.8-.2-.1z"
+							opacity=".2"
+						></path>
+						<path d="M16.1-65.7c-.4-.5-.7-1.2-.7-2.2v.2c0 1 .3 1.7.7 2.2l.1-.1-.1-.1z" opacity=".122"></path>
+						<path d="m49.3-82.7-7.7 4.4.1.1 7.5-4.3c1.1-.6 1.6-1.4 1.6-2.2.1.7-.5 1.5-1.5 2z" opacity=".122"></path>
+						<path
+							d="m19.3-103.8 30 17.1c1 .6 1.5 1.3 1.6 2 0-.8-.5-1.6-1.6-2.2l-30-17.1c-2.2-1.2-3.9-.2-3.9 2.3v.2c0-2.5 1.7-3.5 3.9-2.3z"
+							opacity=".251"
+							fill="#ffffff"
+						></path>
+						<path
+							d="M236.5 51.5v-41c0-.7.1-1.3.2-2 .1-.6.3-1.3.6-1.9.3-.6.6-1.2.9-1.7.4-.5.8-1.1 1.2-1.5.5-.5 1-.9 1.5-1.2.5-.4 1.1-.7 1.7-.9.6-.3 1.2-.4 1.9-.6.6-.1 1.3-.2 2-.2h187c.7 0 1.3.1 2 .2.6.1 1.3.3 1.9.6.6.3 1.2.6 1.7.9.5.4 1.1.8 1.5 1.2.5.5.9 1 1.2 1.5.4.5.7 1.1.9 1.7s.4 1.2.6 1.9c.1.6.2 1.3.2 2v41c0 .7-.1 1.3-.2 2-.1.6-.3 1.3-.6 1.9-.3.6-.6 1.2-.9 1.7s-.8 1.1-1.2 1.5c-.5.5-1 .9-1.5 1.2-.5.4-1.1.7-1.7.9-.6.3-1.2.4-1.9.6s-1.3.2-2 .2h-187c-.7 0-1.3-.1-2-.2-.6-.1-1.3-.3-1.9-.6-.6-.3-1.2-.6-1.7-.9-.5-.4-1.1-.8-1.5-1.2-.5-.5-.9-1-1.2-1.5s-.7-1.1-.9-1.7c-.3-.6-.4-1.2-.6-1.9-.1-.7-.2-1.3-.2-2z"
+							fill="#ffffff"
+						></path>
+						<path
+							d="M236.5 52V10c0-.6.1-1.2.2-1.9.1-.6.3-1.2.5-1.8.2-.6.5-1.1.9-1.6.3-.5.7-1 1.2-1.4.4-.4.9-.8 1.4-1.2.5-.3 1.1-.6 1.6-.9.6-.2 1.2-.4 1.8-.5.6-.1 1.2-.2 1.9-.2h188c.6 0 1.2.1 1.9.2.6.1 1.2.3 1.8.5.6.2 1.1.5 1.6.9.5.3 1 .7 1.4 1.2.4.4.8.9 1.2 1.4s.6 1.1.9 1.6c.2.6.4 1.2.5 1.8.1.6.2 1.2.2 1.9v42c0 .6-.1 1.2-.2 1.9-.1.6-.3 1.2-.5 1.8-.2.6-.5 1.1-.9 1.6-.3.5-.7 1-1.2 1.4-.4.4-.9.8-1.4 1.2-.5.3-1.1.6-1.6.9-.6.2-1.2.4-1.8.5-.6.1-1.2.2-1.9.2H246c-.6 0-1.2-.1-1.9-.2-.6-.1-1.2-.3-1.8-.5s-1.1-.5-1.6-.9c-.5-.3-1-.7-1.4-1.2-.4-.4-.8-.9-1.2-1.4-.3-.5-.6-1.1-.9-1.6-.2-.6-.4-1.2-.5-1.8-.1-.7-.2-1.3-.2-1.9z"
+							fill="none"
+							stroke="#a5a5a5"
+						></path>
+						<path d="M290.2 32.1c0-4.4 3.6-6.5 3.8-6.6-2.1-3-5.3-3.4-6.4-3.4-2.7-.3-5.3 1.6-6.7 1.6-1.4 0-3.5-1.6-5.8-1.5-2.9 0-5.7 1.7-7.2 4.4-3.1 5.4-.8 13.2 2.2 17.5 1.5 2.1 3.2 4.5 5.5 4.4 2.2-.1 3-1.4 5.7-1.4s3.4 1.4 5.7 1.4c2.4 0 3.9-2.1 5.3-4.3 1.7-2.4 2.4-4.8 2.4-4.9.1-.2-4.5-1.9-4.5-7.2z"></path>
+						<path d="M285.8 19.2c1.2-1.5 2-3.5 1.8-5.6-1.7.1-3.9 1.2-5.1 2.7-1.1 1.3-2.1 3.4-1.8 5.4 1.9.1 3.9-1.1 5.1-2.5z"></path>
+						<path d="M325.1 35.1h-6.2l-1.5 4.4h-2.6l5.9-16.3h2.7l5.9 16.3h-2.7l-1.5-4.4zm-5.6-2h4.9l-2.4-7.2h-.1l-2.4 7.2z"></path>
+						<path d="M342 33.6c0 3.7-2 6.1-5 6.1-1.7 0-3-.8-3.7-2.1h-.1v5.9h-2.4V27.6h2.4v2c.7-1.3 2.1-2.1 3.8-2.1 3 0 5 2.4 5 6.1zm-2.5 0c0-2.4-1.2-4-3.1-4-1.9 0-3.1 1.6-3.1 4s1.3 4 3.1 4 3.1-1.6 3.1-4z"></path>
+						<path d="M355.1 33.6c0 3.7-2 6.1-5 6.1-1.7 0-3-.8-3.7-2.1h-.1v5.9h-2.4V27.6h2.4v2c.7-1.3 2.1-2.1 3.8-2.1 3 0 5 2.4 5 6.1zm-2.5 0c0-2.4-1.2-4-3.1-4-1.9 0-3.1 1.6-3.1 4s1.3 4 3.1 4 3.1-1.6 3.1-4z"></path>
+						<path d="M363.8 35c.2 1.6 1.8 2.7 3.9 2.7 2.1 0 3.5-1.1 3.5-2.5 0-1.3-.9-2-3-2.5l-2.1-.5c-3-.7-4.4-2.1-4.4-4.4 0-2.8 2.5-4.7 5.9-4.7 3.4 0 5.8 1.9 5.9 4.7H371c-.1-1.6-1.5-2.6-3.5-2.6s-3.3 1-3.3 2.4c0 1.2.9 1.8 3 2.4l1.8.4c3.3.8 4.7 2.1 4.7 4.5 0 3.1-2.4 5-6.3 5-3.6 0-6.1-1.9-6.2-4.8h2.6z"></path>
+						<path d="M379.1 24.8v2.8h2.3v1.9h-2.3v6.6c0 1 .5 1.5 1.4 1.5.2 0 .6 0 .8-.1v1.9c-.3.1-.8.1-1.4.1-2.4 0-3.3-.9-3.3-3.2v-6.8h-1.7v-1.9h1.7v-2.8h2.5z"></path>
+						<path d="M382.7 33.6c0-3.7 2.2-6.1 5.6-6.1 3.4 0 5.6 2.4 5.6 6.1 0 3.8-2.2 6.1-5.6 6.1s-5.6-2.4-5.6-6.1zm8.8 0c0-2.6-1.2-4.1-3.2-4.1s-3.2 1.5-3.2 4.1c0 2.6 1.2 4.1 3.2 4.1s3.2-1.5 3.2-4.1z"></path>
+						<path d="M396 27.6h2.3v2h.1c.4-1.3 1.5-2.1 2.9-2.1.4 0 .6 0 .8.1v2.3c-.2-.1-.6-.1-1.1-.1-1.6 0-2.5 1.1-2.5 2.7v7.1H396v-12z"></path>
+						<path d="M413.3 36c-.3 2.2-2.4 3.6-5.1 3.6-3.5 0-5.6-2.3-5.6-6s2.2-6.2 5.5-6.2 5.4 2.3 5.4 5.9v.8h-8.4v.1c0 2 1.3 3.4 3.2 3.4 1.4 0 2.4-.6 2.7-1.7h2.3zm-8.2-3.5h5.9c-.1-1.8-1.2-3-2.9-3s-2.9 1.2-3 3z"></path>
+						<path d="M424.2-115.6H251c-1 0-2.1.1-3.1.3-1 .2-2 .5-2.9 1-.9.5-1.8 1.1-2.5 1.8s-1.3 1.6-1.8 2.5c-.5.9-.8 1.9-1 2.9-.2 1-.2 2-.3 3.1v38.4c0 1 .1 2.1.3 3.1.2 1 .5 2 1 2.9.5.9 1.1 1.8 1.8 2.5s1.6 1.3 2.5 1.8c.9.5 1.9.8 2.9 1 1 .2 2.1.2 3.1.3h176.3c1 0 2.1-.1 3.1-.3 1-.2 2-.5 2.9-1 .9-.5 1.8-1.1 2.5-1.8s1.3-1.6 1.8-2.5c.5-.9.8-1.9 1-2.9.2-1 .2-2 .3-3.1V-104c0-1-.1-2.1-.3-3.1-.2-1-.5-2-1-2.9-.5-.9-1.1-1.8-1.8-2.5s-1.6-1.3-2.5-1.8c-.9-.5-1.9-.8-2.9-1-1-.2-2-.2-3.1-.3h-3.1z"></path>
+						<path
+							d="M252.4-55.4H251c-.9 0-1.9-.1-2.9-.3-.9-.2-1.8-.4-2.5-.8-.8-.4-1.5-.9-2.1-1.6-.6-.6-1.2-1.3-1.6-2.1-.4-.8-.7-1.6-.8-2.5-.2-1-.2-2.1-.3-2.9v-38.3c0-.8.1-1.9.3-2.9.2-.9.4-1.8.8-2.6s.9-1.5 1.6-2.1c.6-.6 1.4-1.2 2.2-1.6.8-.4 1.6-.7 2.5-.8 1-.2 2.1-.2 2.9-.3h176.2c.8 0 1.8.1 2.9.2.9.2 1.8.4 2.6.8s1.5.9 2.1 1.6c.6.6 1.2 1.3 1.6 2.2.4.8.7 1.6.8 2.5.2 1 .2 2 .3 2.9v38.4c0 .9-.1 1.9-.3 2.8-.2.9-.4 1.8-.8 2.6s-.9 1.5-1.6 2.1c-.6.6-1.4 1.2-2.2 1.6-.8.4-1.6.7-2.6.8-1 .2-2 .2-2.9.3H252.4z"
+							fill="#ffffff"
+						></path>
+						<path d="M277.9-85c0-4.4 3.6-6.5 3.8-6.6-2.1-3-5.3-3.4-6.4-3.4-2.7-.3-5.3 1.6-6.7 1.6-1.4 0-3.5-1.6-5.8-1.5-2.9 0-5.7 1.7-7.2 4.4-3.1 5.4-.8 13.2 2.2 17.5 1.5 2.1 3.2 4.5 5.5 4.4 2.2-.1 3-1.4 5.7-1.4s3.4 1.4 5.7 1.4c2.4 0 3.9-2.1 5.3-4.3 1.7-2.4 2.4-4.8 2.4-4.9.1-.2-4.5-1.9-4.5-7.2z"></path>
+						<path d="M273.5-97.9c1.2-1.5 2-3.5 1.8-5.6-1.7.1-3.9 1.2-5.1 2.7-1.1 1.3-2.1 3.4-1.8 5.4 1.9.1 3.9-1.1 5.1-2.5z"></path>
+						<path d="M297.8-76.7h-3.9v5.4c1.6 0 3.3-.1 4.8-.4l.1 1.2c-1.2.2-3.4.4-5.5.4h-.8v-7.7h5.3v1.1zm3.2 2.2h1.7v1.2H301v5.9h-1.4v-11.5h1.4v4.4z"></path>
+						<path d="M314.1-73.3v1.1h-4.4v2.2h-1.4v-2.2h-4.6v-1.1h10.4zm-9.2-3c0-1.3 1.5-2.3 4-2.3 2.4 0 4 1 4 2.3 0 1.4-1.6 2.3-4 2.3s-4-1-4-2.3zm1.5 7.4h6.5v1.1H305V-71h1.4v2.1zm0-7.4c0 .7 1.1 1.2 2.5 1.2 1.5 0 2.5-.5 2.5-1.2 0-.8-1-1.2-2.5-1.2-1.4 0-2.5.4-2.5 1.2z"></path>
+						<path d="M325.8-70v1.1h-10.4V-70h4.5v-1.9h-3.1v-3.8h6.3v-1.4h-6.3v-1.1h7.8v3.7h-6.3v1.5h6.6v1.1h-3.5v1.9h4.4z"></path>
+						<path d="M337.5-69h-10.4v-1.1h10.4v1.1zm-1.4-7.8h-6.3v3.2h6.4v1.1h-7.8V-78h7.7v1.2z"></path>
+						<path d="M348.9-75h-6.7v-1.1h6.7v1.1zm-3.3 5.4c-1.6 0-2.7-1-2.7-2.4s1.2-2.4 2.7-2.4c1.6 0 2.8 1 2.8 2.4s-1.2 2.4-2.8 2.4zm1.7-7.4h-3.5v-1.2h3.5v1.2zm-1.7 6.2c.8 0 1.4-.5 1.4-1.3s-.6-1.3-1.4-1.3c-.8 0-1.4.5-1.4 1.3s.6 1.3 1.4 1.3zm5.6-3.4h1.6v1.2h-1.6v5.6h-1.4v-11.5h1.4v4.7z"></path>
+						<path d="m354.8-69.5-.8-1.1c2.7-1.3 4.3-3.6 4.5-5.9h-3.9v-1.2h5.4c-.1 3.4-1.7 6.3-5.2 8.2zm8.6 2.1H362v-11.5h1.4v11.5z"></path>
+						<path d="M302.7-90.1h-6.2l-1.5 4.4h-2.6l5.9-16.3h2.7l5.9 16.3h-2.7l-1.5-4.4zm-5.5-2.1h4.9l-2.4-7.2h-.1l-2.4 7.2z"></path>
+						<path d="M319.7-91.7c0 3.7-2 6.1-5 6.1-1.7 0-3-.8-3.7-2.1h-.1v5.9h-2.4v-15.8h2.4v2c.7-1.3 2.1-2.1 3.8-2.1 3 0 5 2.4 5 6zm-2.5 0c0-2.4-1.2-4-3.1-4-1.9 0-3.1 1.6-3.1 4s1.3 4 3.1 4 3.1-1.5 3.1-4z"></path>
+						<path d="M332.8-91.7c0 3.7-2 6.1-5 6.1-1.7 0-3-.8-3.7-2.1h-.1v5.9h-2.4v-15.8h2.4v2c.7-1.3 2.1-2.1 3.8-2.1 3 0 5 2.4 5 6zm-2.5 0c0-2.4-1.2-4-3.1-4-1.9 0-3.1 1.6-3.1 4s1.3 4 3.1 4 3.1-1.5 3.1-4z"></path>
+						<path d="M341.5-90.3c.2 1.6 1.8 2.7 3.9 2.7 2.1 0 3.5-1.1 3.5-2.5 0-1.3-.9-2-3-2.5l-2.1-.5c-3-.7-4.4-2.1-4.4-4.4 0-2.8 2.5-4.7 5.9-4.7 3.4 0 5.8 1.9 5.9 4.7h-2.5c-.1-1.6-1.5-2.6-3.5-2.6s-3.3 1-3.3 2.4c0 1.2.9 1.8 3 2.4l1.8.4c3.3.8 4.7 2.1 4.7 4.5 0 3.1-2.4 5-6.3 5-3.6 0-6.1-1.9-6.2-4.8h2.6z"></path>
+						<path d="M356.8-100.4v2.8h2.3v1.9h-2.3v6.6c0 1 .5 1.5 1.4 1.5.2 0 .6 0 .8-.1v1.9c-.3.1-.8.1-1.4.1-2.4 0-3.3-.9-3.3-3.2v-6.8h-1.7v-1.9h1.7v-2.8h2.5z"></path>
+						<path d="M360.4-91.7c0-3.7 2.2-6.1 5.6-6.1 3.4 0 5.6 2.4 5.6 6.1 0 3.8-2.2 6.1-5.6 6.1s-5.6-2.3-5.6-6.1zm8.8 0c0-2.6-1.2-4.1-3.2-4.1s-3.2 1.5-3.2 4.1c0 2.6 1.2 4.1 3.2 4.1s3.2-1.5 3.2-4.1z"></path>
+						<path d="M373.7-97.6h2.3v2h.1c.4-1.3 1.5-2.1 2.9-2.1.4 0 .6 0 .8.1v2.3c-.2-.1-.6-.1-1.1-.1-1.6 0-2.5 1.1-2.5 2.7v7.1h-2.4v-12z"></path>
+						<path d="M391-89.2c-.3 2.2-2.4 3.6-5.1 3.6-3.5 0-5.6-2.3-5.6-6s2.2-6.2 5.5-6.2 5.4 2.3 5.4 5.9v.8h-8.4v.1c0 2 1.3 3.4 3.2 3.4 1.4 0 2.4-.6 2.7-1.7h2.3zm-8.2-3.6h5.9c-.1-1.8-1.2-3-2.9-3s-2.9 1.3-3 3z"></path>
+						<path d="M401-101.7h1.9v15.3H401v-7.7h-1.4c-.2 2.7-1.3 5.1-3.5 5.1-2.5 0-3.5-2.8-3.5-5.9 0-3 1.1-5.8 3.5-5.8 2.2 0 3.3 2.2 3.5 4.8h1.4v-5.8zm-3.2 6.7c0-2.2-.5-4-1.7-4-1.1 0-1.6 1.7-1.6 4s.5 3.9 1.6 3.9c1.2.1 1.7-1.6 1.7-3.9zm8.5-7.1v16.4h-1.9v-16.4h1.9z"></path>
+						<path d="M414.9-98.2c0 3.3 1.5 5.7 4 7.1l-1.2 1.5c-1.5-.9-3.4-2.7-3.9-4.4-.4 1.8-2.3 3.9-3.9 4.8l-1.3-1.5c2.7-1.6 4.1-4.2 4.1-7.5v-2.5h2.1v2.5zm7.6 12.5h-2v-9h-3.3v-1.7h3.3v-5.6h2v16.3z"></path>
+					</g>
+				</svg>
+			</Box>
+		</a>
+	);
+}
 
 const Box = styled.div`
-  width: 200px;
-  @media ${(props) => props.theme.breakpoints.md} {
-    width: 130px;
-  }
-  @media ${(props) => props.theme.breakpoints.sm} {
-    width: 100px;
-  }
+	width: 200px;
+	@media ${(props) => props.theme.breakpoints.md} {
+		width: 130px;
+	}
+	@media ${(props) => props.theme.breakpoints.sm} {
+		width: 100px;
+	}
 `;
-
-const GoogleApp = () => {
-  return (
-    <a
-      href="https://play.google.com/store/apps/details?id=com.stockeeper.stockeeperapp"
-      target="_blank"
-      rel="noreferrer"
-    >
-      <Box>
-        <svg
-          preserveAspectRatio="xMidYMid meet"
-          data-bbox="0.5 -115.6 443 177.1"
-          viewBox="0 0 208 62"
-          xmlns="http://www.w3.org/2000/svg"
-          data-type="ugc"
-          role="presentation"
-          aria-hidden="true"
-          aria-labelledby="svgcid-i4ic11-ni6gtn"
-        >
-          <title id="svgcid-i4ic11-ni6gtn"></title>
-          <g>
-            <path
-              d="M16.1 49.9c-.4-.5-.7-1.2-.7-2.2v.2c0 1 .3 1.7.7 2.2l.1-.1-.1-.1z"
-              opacity=".122"
-            ></path>
-            <path
-              d="M.5 51.5v-41c0-.7.1-1.3.2-2 .1-.6.3-1.2.6-1.8s.6-1.2.9-1.7c.4-.6.8-1.1 1.2-1.6.5-.5 1-.9 1.5-1.2.5-.4 1.1-.7 1.7-.9C7.3 1 7.9.8 8.5.7s1.3-.2 2-.2h187c.7 0 1.3.1 2 .2.6.1 1.3.3 1.9.6.6.3 1.2.6 1.7.9.5.4 1.1.8 1.5 1.2.5.5.9 1 1.2 1.5.4.5.7 1.1.9 1.7.3.6.4 1.2.6 1.9.1.6.2 1.3.2 2v41c0 .7-.1 1.3-.2 2-.1.6-.3 1.3-.6 1.9-.3.6-.6 1.2-.9 1.7s-.8 1.1-1.2 1.5c-.5.5-1 .9-1.5 1.2-.5.4-1.1.7-1.7.9-.6.3-1.2.4-1.9.6s-1.3.2-2 .2h-187c-.7 0-1.3-.1-2-.2s-1.2-.3-1.8-.6c-.6-.3-1.2-.6-1.7-.9-.5-.4-1.1-.8-1.5-1.2-.5-.5-.9-1-1.2-1.5s-.7-1.1-.9-1.7c-.3-.6-.4-1.2-.6-1.9-.2-.7-.3-1.3-.3-2z"
-              fill="#ffffff"
-            ></path>
-            <path
-              d="M.5 52V10c0-.6.1-1.2.2-1.9.1-.6.3-1.2.5-1.7.2-.6.5-1.1.9-1.6s.7-1 1.2-1.4.9-.8 1.4-1.2 1.1-.6 1.6-.9C6.9 1 7.5.8 8.1.7S9.4.5 10 .5h188c.6 0 1.2.1 1.9.2s1.2.3 1.8.5c.6.2 1.1.5 1.6.9s1 .7 1.4 1.2c.4.4.8.9 1.2 1.4s.6 1.1.9 1.6c.2.6.4 1.2.5 1.8.1.6.2 1.2.2 1.9v42c0 .6-.1 1.2-.2 1.9s-.3 1.2-.5 1.8c-.2.6-.5 1.1-.9 1.6-.3.5-.7 1-1.2 1.4-.4.4-.9.8-1.4 1.2-.5.3-1.1.6-1.6.9-.6.2-1.2.4-1.8.5-.6.1-1.2.2-1.9.2H10c-.6 0-1.2-.1-1.9-.2-.6-.1-1.2-.3-1.8-.5s-1.1-.5-1.6-.9c-.5-.3-1-.7-1.4-1.2-.4-.4-.8-.9-1.2-1.4s-.6-1.1-.9-1.6c-.2-.6-.4-1.2-.5-1.8-.1-.7-.2-1.3-.2-1.9z"
-              fill="none"
-              stroke="#a5a5a5"
-            ></path>
-            <path d="M164.2 39.8h2.9V20.6h-2.9v19.2zM190 27.5l-3.3 8.4h-.1l-3.4-8.4h-3.1l5.1 11.7-2.9 6.5h3l7.9-18.2H190zm-16.3 10.2c-.9 0-2.3-.5-2.3-1.6 0-1.5 1.6-2.1 3-2.1 1.3 0 1.9.3 2.6.6-.1 1.8-1.6 3.1-3.3 3.1zm.4-10.6c-2.1 0-4.2.9-5.1 2.9l2.6 1.1c.5-1.1 1.6-1.4 2.6-1.4 1.5 0 3 .9 3 2.5v.2c-.5-.3-1.6-.7-3-.7-2.8 0-5.6 1.5-5.6 4.3 0 2.6 2.3 4.2 4.8 4.2 1.9 0 3-.9 3.7-1.9h.1v1.5h2.8v-7.4c-.1-3.4-2.6-5.3-5.9-5.3zm-17.8 2.8h-4.1v-6.6h4.1c2.1 0 3.4 1.8 3.4 3.3 0 1.5-1.2 3.3-3.4 3.3zm-.1-9.3h-6.9v19.3h2.9v-7.3h4c3.2 0 6.3-2.3 6.3-6 .1-3.7-3.1-6-6.3-6zm-37.5 17.1c-2 0-3.6-1.7-3.6-3.9 0-2.3 1.7-4 3.6-4 2 0 3.5 1.7 3.5 4 0 2.2-1.6 3.9-3.5 3.9zm3.3-9.1c-.7-.8-2-1.5-3.5-1.5-3.3 0-6.3 2.9-6.3 6.6 0 3.7 3 6.5 6.3 6.5 1.6 0 2.8-.7 3.4-1.5h.1v.9c0 2.5-1.3 3.8-3.5 3.8-1.8 0-2.9-1.3-3.3-2.3l-2.5 1c.7 1.7 2.6 3.9 5.8 3.9 3.4 0 6.2-2 6.2-6.8V27.5H122v1.1zm4.7 11.2h2.9V20.6h-2.9v19.2zm7.1-6.3c-.1-2.5 2-3.8 3.4-3.8 1.1 0 2.1.6 2.4 1.4l-5.8 2.4zm9-2.2c-.5-1.5-2.2-4.2-5.6-4.2-3.4 0-6.2 2.7-6.2 6.6 0 3.7 2.8 6.6 6.5 6.6 3 0 4.7-1.8 5.5-2.9l-2.2-1.5c-.7 1.1-1.8 1.8-3.2 1.8-1.5 0-2.5-.7-3.2-2l8.8-3.6-.4-.8zM73 29.1v2.8h6.7c-.2 1.6-.7 2.7-1.5 3.5-1 1-2.5 2-5.1 2-4.1 0-7.3-3.3-7.3-7.4 0-4.1 3.2-7.4 7.3-7.4 2.2 0 3.8.9 5 2l2-2c-1.7-1.6-3.9-2.8-7-2.8-5.6 0-10.3 4.6-10.3 10.2s4.7 10.2 10.3 10.2c3 0 5.3-1 7.1-2.9 1.8-1.8 2.4-4.4 2.4-6.5 0-.6 0-1.2-.1-1.7H73zm17 8.6c-2 0-3.7-1.6-3.7-4s1.7-4 3.7-4 3.7 1.6 3.7 4c0 2.3-1.7 4-3.7 4zm0-10.6c-3.6 0-6.6 2.8-6.6 6.6s3 6.6 6.6 6.6c3.6 0 6.6-2.8 6.6-6.6s-3-6.6-6.6-6.6zm14.4 10.6c-2 0-3.7-1.6-3.7-4s1.7-4 3.7-4 3.7 1.6 3.7 4c0 2.3-1.7 4-3.7 4zm0-10.6c-3.6 0-6.6 2.8-6.6 6.6s3 6.6 6.6 6.6c3.6 0 6.6-2.8 6.6-6.6-.1-3.8-3-6.6-6.6-6.6z"></path>
-            <g transform="translate(25.366 20.986)">
-              <linearGradient
-                gradientTransform="matrix(19.9196 0 0 -38.628 6492.775 7042.514)"
-                y2="181.09"
-                x2="-324.559"
-                y1="180.421"
-                x1="-323.261"
-                gradientUnits="userSpaceOnUse"
-                id="svgcid-a3x0nd5iltcx"
-              >
-                <stop offset="0" stopColor="#00a0ff"></stop>
-                <stop offset=".007" stopColor="#00a1ff"></stop>
-                <stop offset=".26" stopColor="#00beff"></stop>
-                <stop offset=".512" stopColor="#00d2ff"></stop>
-                <stop offset=".76" stopColor="#00dfff"></stop>
-                <stop offset="1" stopColor="#00e3ff"></stop>
-              </linearGradient>
-              <path
-                d="M-9.9-9.3c-.5.6-.8 1.4-.7 2.2V27c0 .8.2 1.6.7 2.2l.1.1L9.3 10.1v-.4L-9.8-9.4l-.1.1z"
-                fill="url(#svgcid-a3x0nd5iltcx)"
-              ></path>
-              <linearGradient
-                gradientTransform="matrix(15.661 0 0 -13.181 5054.71 2158.078)"
-                y2="162.976"
-                x2="-323.468"
-                y1="162.976"
-                x1="-321.087"
-                gradientUnits="userSpaceOnUse"
-                id="svgcid--g2g3vz-5nrlpi"
-              >
-                <stop offset="0" stopColor="#ffe000"></stop>
-                <stop offset=".409" stopColor="#ffbd00"></stop>
-                <stop offset=".775" stopColor="#ffa500"></stop>
-                <stop offset="1" stopColor="#ff9c00"></stop>
-              </linearGradient>
-              <path
-                d="m15.7 16.5-6.4-6.4v-.4l6.4-6.4.1.1 7.5 4.3c2.2 1.2 2.2 3.2 0 4.4l-7.5 4.3-.1.1z"
-                fill="url(#svgcid--g2g3vz-5nrlpi)"
-              ></path>
-              <linearGradient
-                gradientTransform="matrix(25.708 0 0 -19.8124 8475.259 3440.851)"
-                y2="167.706"
-                x2="-325.849"
-                y1="165.936"
-                x1="-324.486"
-                gradientUnits="userSpaceOnUse"
-                id="svgcid-11ngs6-9ko6u8"
-              >
-                <stop offset="0" stopColor="#ff3a44"></stop>
-                <stop offset="1" stopColor="#c31162"></stop>
-              </linearGradient>
-              <path
-                d="M15.8 16.4 9.3 9.9-9.9 29.1c.7.8 1.9.8 3.2.1l22.5-12.8"
-                fill="url(#svgcid-11ngs6-9ko6u8)"
-              ></path>
-              <linearGradient
-                gradientTransform="matrix(25.707 0 0 -19.8124 8474.918 3421.039)"
-                y2="164.979"
-                x2="-335.604"
-                y1="164.189"
-                x1="-336.213"
-                gradientUnits="userSpaceOnUse"
-                id="svgcid-4cpanjacuzsy"
-              >
-                <stop offset="0" stopColor="#32a071"></stop>
-                <stop offset=".068" stopColor="#2da771"></stop>
-                <stop offset=".476" stopColor="#15cf74"></stop>
-                <stop offset=".801" stopColor="#06e775"></stop>
-                <stop offset="1" stopColor="#00f076"></stop>
-              </linearGradient>
-              <path
-                d="M15.8 3.4-6.7-9.4c-1.3-.8-2.5-.7-3.2.1L9.3 9.9l6.5-6.5z"
-                fill="url(#svgcid-4cpanjacuzsy)"
-              ></path>
-              <path
-                d="m15.7 16.279-22.4 12.7c-.9.7-2.2.7-3.1 0l-.1.1.1.1c.9.7 2.2.7 3.1 0l22.5-12.8-.1-.1z"
-                opacity=".2"
-              ></path>
-              <path
-                d="M-9.9 28.879c-.5-.6-.8-1.4-.7-2.2v.2c0 .8.2 1.6.7 2.2l.1-.1-.1-.1z"
-                opacity=".12"
-              ></path>
-              <path
-                d="m23.4 11.9-7.7 4.4.1.1 7.5-4.3c.9-.4 1.5-1.2 1.6-2.2-.1.9-.7 1.6-1.5 2z"
-                opacity=".12"
-              ></path>
-              <path
-                d="m-6.7-9.2 30 17.1c.8.4 1.4 1.1 1.6 2-.1-1-.7-1.8-1.6-2.2l-30-17.1c-2.2-1.2-3.9-.2-3.9 2.3v.2c0-2.5 1.8-3.5 3.9-2.3z"
-                opacity=".25"
-                fill="#ffffff"
-              ></path>
-            </g>
-            <path
-              d="M.5-64.1v-41c0-.7.1-1.3.2-2 .1-.6.3-1.3.6-1.9s.6-1.2.9-1.7c.4-.5.8-1.1 1.2-1.5.5-.5 1-.9 1.5-1.2.5-.4 1.1-.7 1.7-.9s1.2-.4 1.9-.6 1.3-.2 2-.2h187c.7 0 1.3.1 2 .2.6.1 1.3.3 1.9.6.6.3 1.2.6 1.7.9.5.4 1.1.8 1.5 1.2.5.5.9 1 1.2 1.5.4.5.7 1.1.9 1.7.3.6.4 1.2.6 1.9.1.6.2 1.3.2 2v41c0 .7-.1 1.3-.2 2-.1.6-.3 1.3-.6 1.9-.3.6-.6 1.2-.9 1.7s-.8 1.1-1.2 1.5c-.5.5-1 .9-1.5 1.2-.5.4-1.1.7-1.7.9-.6.3-1.2.4-1.9.6s-1.3.2-2 .2h-187c-.7 0-1.3-.1-2-.2s-1.3-.3-1.9-.6c-.6-.3-1.2-.6-1.7-.9-.5-.4-1.1-.8-1.5-1.2-.5-.5-.9-1-1.2-1.5s-.7-1.1-.9-1.7c-.3-.6-.4-1.2-.6-1.9-.1-.7-.2-1.3-.2-2z"
-              fill="#ffffff"
-            ></path>
-            <path
-              d="M.5-63.6v-42c0-.6.1-1.2.2-1.9s.3-1.2.5-1.8.5-1.1.9-1.6.7-1 1.2-1.4.9-.8 1.4-1.2c.5-.3 1.1-.6 1.6-.9.6-.2 1.2-.4 1.8-.5.6-.1 1.2-.2 1.9-.2h188c.6 0 1.2.1 1.9.2.6.1 1.2.3 1.8.5.6.2 1.1.5 1.6.9.5.3 1 .7 1.4 1.2.4.4.8.9 1.2 1.4s.6 1.1.9 1.6.4 1.2.5 1.8.2 1.2.2 1.9v42c0 .6-.1 1.2-.2 1.9s-.3 1.2-.5 1.8c-.2.6-.5 1.1-.9 1.6-.3.5-.7 1-1.2 1.4-.4.4-.9.8-1.4 1.2-.5.3-1.1.6-1.6.9-.6.2-1.2.4-1.8.5-.6.1-1.2.2-1.9.2H10c-.6 0-1.2-.1-1.9-.2-.6-.1-1.2-.3-1.8-.5s-1.1-.5-1.6-.9c-.5-.3-1-.7-1.4-1.2-.4-.4-.8-.9-1.2-1.4s-.6-1.1-.9-1.6c-.2-.6-.4-1.2-.5-1.8-.1-.6-.2-1.2-.2-1.9z"
-              fill="none"
-              stroke="#a5a5a5"
-            ></path>
-            <path d="M73.1-99.7c0 1.3-.4 2.3-1.1 3.1-.9.9-2 1.4-3.4 1.4-1.3 0-2.5-.5-3.4-1.4-.9-.9-1.4-2.1-1.4-3.4 0-1.4.5-2.5 1.4-3.4.9-.9 2.1-1.4 3.4-1.4.7 0 1.3.1 1.9.4.6.3 1.1.6 1.4 1l-.8.8c-.6-.7-1.5-1.1-2.5-1.1s-1.8.3-2.5 1-1.1 1.6-1.1 2.7.4 2 1.1 2.7c.7.7 1.5 1 2.5 1s1.9-.3 2.6-1c.4-.4.7-1.1.8-1.9h-3.4v-1.1H73c0 .1.1.3.1.6z"></path>
-            <path
-              d="M73.1-99.7c0 1.3-.4 2.3-1.1 3.1-.9.9-2 1.4-3.4 1.4-1.3 0-2.5-.5-3.4-1.4-.9-.9-1.4-2.1-1.4-3.4 0-1.4.5-2.5 1.4-3.4.9-.9 2.1-1.4 3.4-1.4.7 0 1.3.1 1.9.4.6.3 1.1.6 1.4 1l-.8.8c-.6-.7-1.5-1.1-2.5-1.1s-1.8.3-2.5 1-1.1 1.6-1.1 2.7.4 2 1.1 2.7c.7.7 1.5 1 2.5 1s1.9-.3 2.6-1c.4-.4.7-1.1.8-1.9h-3.4v-1.1H73c0 .1.1.3.1.6z"
-              fill="none"
-              stroke="#ffffff"
-              strokeWidth=".2"
-              strokeMiterlimit="10"
-            ></path>
-            <path d="M80.2-103.6H76v2.9h3.7v1.2H76v2.9h4.2v1.1h-5.4v-9.2h5.4v1.1z"></path>
-            <path
-              fill="none"
-              stroke="#ffffff"
-              strokeWidth=".2"
-              strokeMiterlimit="10"
-              d="M80.2-103.6H76v2.9h3.7v1.2H76v2.9h4.2v1.1h-5.4v-9.2h5.4v1.1z"
-            ></path>
-            <path d="M85.2-95.5H84v-8.1h-2.6v-1.1h6.4v1.1h-2.6v8.1z"></path>
-            <path
-              fill="none"
-              stroke="#ffffff"
-              strokeWidth=".2"
-              strokeMiterlimit="10"
-              d="M85.2-95.5H84v-8.1h-2.6v-1.1h6.4v1.1h-2.6v8.1z"
-            ></path>
-            <path d="M93.5-104.7v9.2h-1.2v-9.2h1.2z"></path>
-            <path
-              fill="none"
-              stroke="#ffffff"
-              strokeWidth=".2"
-              strokeMiterlimit="10"
-              d="M93.5-104.7v9.2h-1.2v-9.2h1.2z"
-            ></path>
-            <path d="M98.8-95.5h-1.2v-8.1H95v-1.1h6.4v1.1h-2.6v8.1z"></path>
-            <path
-              fill="none"
-              stroke="#ffffff"
-              strokeWidth=".2"
-              strokeMiterlimit="10"
-              d="M98.8-95.5h-1.2v-8.1H95v-1.1h6.4v1.1h-2.6v8.1z"
-            ></path>
-            <path d="M107.5-97.4c.7.7 1.5 1 2.5 1s1.8-.3 2.5-1 1-1.6 1-2.7-.3-2-1-2.7c-.7-.7-1.5-1-2.5-1s-1.8.3-2.5 1-1 1.6-1 2.7.3 2 1 2.7zm5.9.7c-.9.9-2 1.4-3.4 1.4s-2.5-.5-3.4-1.4c-.9-.9-1.4-2.1-1.4-3.4 0-1.4.5-2.5 1.4-3.4.9-.9 2-1.4 3.4-1.4 1.3 0 2.5.5 3.4 1.4.9.9 1.4 2.1 1.4 3.4 0 1.3-.5 2.5-1.4 3.4z"></path>
-            <path
-              d="M107.5-97.4c.7.7 1.5 1 2.5 1s1.8-.3 2.5-1 1-1.6 1-2.7-.3-2-1-2.7c-.7-.7-1.5-1-2.5-1s-1.8.3-2.5 1-1 1.6-1 2.7.3 2 1 2.7zm5.9.7c-.9.9-2 1.4-3.4 1.4s-2.5-.5-3.4-1.4c-.9-.9-1.4-2.1-1.4-3.4 0-1.4.5-2.5 1.4-3.4.9-.9 2-1.4 3.4-1.4 1.3 0 2.5.5 3.4 1.4.9.9 1.4 2.1 1.4 3.4 0 1.3-.5 2.5-1.4 3.4z"
-              fill="none"
-              stroke="#ffffff"
-              strokeWidth=".2"
-              strokeMiterlimit="10"
-            ></path>
-            <path d="M116.4-95.5v-9.2h1.5l4.5 7.2v-7.2h1.2v9.2h-1.3l-4.7-7.5v7.5h-1.2z"></path>
-            <path
-              fill="none"
-              stroke="#ffffff"
-              strokeWidth=".2"
-              strokeMiterlimit="10"
-              d="M116.4-95.5v-9.2h1.5l4.5 7.2v-7.2h1.2v9.2h-1.3l-4.7-7.5v7.5h-1.2z"
-            ></path>
-            <path d="M164.8-69.3h2.9v-19.3h-2.9v19.3zm25.9-12.3-3.3 8.4h-.1l-3.4-8.4h-3.1l5.1 11.7-2.9 6.5h3l7.9-18.2h-3.2zm-16.3 10.1c-.9 0-2.3-.5-2.3-1.6 0-1.5 1.6-2.1 3-2.1 1.3 0 1.9.3 2.6.6-.2 1.9-1.7 3.1-3.3 3.1zm.3-10.5c-2.1 0-4.2.9-5.1 2.9l2.6 1.1c.5-1.1 1.6-1.4 2.6-1.4 1.5 0 3 .9 3 2.5v.2c-.5-.3-1.6-.7-3-.7-2.8 0-5.6 1.5-5.6 4.3 0 2.6 2.3 4.2 4.8 4.2 1.9 0 3-.9 3.7-1.9h.1v1.5h2.8v-7.4c-.1-3.4-2.6-5.3-5.9-5.3zm-17.8 2.7h-4.1v-6.6h4.1c2.1 0 3.4 1.8 3.4 3.3 0 1.5-1.2 3.3-3.4 3.3zm0-9.2H150v19.3h2.9v-7.3h4c3.2 0 6.3-2.3 6.3-6s-3.2-6-6.3-6zm-37.6 17c-2 0-3.6-1.7-3.6-3.9 0-2.3 1.7-4 3.6-4 2 0 3.5 1.7 3.5 4s-1.5 3.9-3.5 3.9zm3.3-9c-.7-.8-2-1.5-3.5-1.5-3.3 0-6.3 2.9-6.3 6.6 0 3.7 3 6.5 6.3 6.5 1.6 0 2.8-.7 3.4-1.5h.1v.9c0 2.5-1.3 3.8-3.5 3.8-1.8 0-2.9-1.3-3.3-2.3l-2.5 1c.7 1.7 2.6 3.9 5.8 3.9 3.4 0 6.2-2 6.2-6.8v-11.8h-2.7v1.2zm4.7 11.2h2.9v-19.3h-2.9v19.3zm7.1-6.3c-.1-2.5 2-3.8 3.4-3.8 1.1 0 2.1.6 2.4 1.4l-5.8 2.4zm9-2.2c-.5-1.5-2.2-4.2-5.6-4.2-3.4 0-6.2 2.7-6.2 6.6 0 3.7 2.8 6.6 6.5 6.6 3 0 4.7-1.8 5.5-2.9l-2.2-1.5c-.7 1.1-1.8 1.8-3.2 1.8-1.5 0-2.5-.7-3.2-2l8.8-3.6-.4-.8zM73.6-80v2.8h6.7c-.2 1.6-.7 2.7-1.5 3.5-1 1-2.5 2-5.1 2-4.1 0-7.3-3.3-7.3-7.4 0-4.1 3.2-7.4 7.3-7.4 2.2 0 3.8.9 5 2l2-2c-1.7-1.6-3.9-2.8-7-2.8-5.6 0-10.3 4.6-10.3 10.2S68-68.9 73.6-68.9c3 0 5.3-1 7.1-2.9 1.8-1.8 2.4-4.4 2.4-6.5 0-.6 0-1.2-.1-1.7h-9.4zm17 8.5c-2 0-3.7-1.6-3.7-4s1.7-4 3.7-4 3.7 1.6 3.7 4-1.7 4-3.7 4zm0-10.5C87-82 84-79.2 84-75.4s3 6.6 6.6 6.6c3.6 0 6.6-2.8 6.6-6.6S94.3-82 90.6-82zM105-71.5c-2 0-3.7-1.6-3.7-4s1.7-4 3.7-4 3.7 1.6 3.7 4-1.7 4-3.7 4zm0-10.5c-3.6 0-6.6 2.8-6.6 6.6s3 6.6 6.6 6.6c3.6 0 6.6-2.8 6.6-6.6s-3-6.6-6.6-6.6z"></path>
-            <linearGradient
-              gradientTransform="matrix(1 0 0 -1 -4.607 -175.041)"
-              y2="-191.279"
-              x2="121.728"
-              y1="-217.136"
-              x1="147.586"
-              gradientUnits="userSpaceOnUse"
-              id="svgcid--p8co1pbjv1nx"
-            >
-              <stop offset="0" stopColor="#00a0ff"></stop>
-              <stop offset=".007" stopColor="#00a1ff"></stop>
-              <stop offset=".26" stopColor="#00beff"></stop>
-              <stop offset=".512" stopColor="#00d2ff"></stop>
-              <stop offset=".76" stopColor="#00dfff"></stop>
-              <stop offset="1" stopColor="#00e3ff"></stop>
-            </linearGradient>
-            <path
-              d="M16.1-103.9c-.4.5-.7 1.2-.7 2.2v34.1c0 1 .3 1.7.7 2.2l.1.1 19.1-19.1v-.4L16.2-104l-.1.1z"
-              fill="url(#svgcid--p8co1pbjv1nx)"
-            ></path>
-            <linearGradient
-              gradientTransform="matrix(1 0 0 -1 2.38 -170.576)"
-              y2="-85.885"
-              x2="12.467"
-              y1="-85.885"
-              x1="49.748"
-              gradientUnits="userSpaceOnUse"
-              id="svgcid--qj0fwh-s7c4e8"
-            >
-              <stop offset="0" stopColor="#ffe000"></stop>
-              <stop offset=".409" stopColor="#ffbd00"></stop>
-              <stop offset=".775" stopColor="#ffa500"></stop>
-              <stop offset="1" stopColor="#ff9c00"></stop>
-            </linearGradient>
-            <path
-              d="m41.6-78.1-6.4-6.4v-.4l6.4-6.4.1.1 7.5 4.3c2.2 1.2 2.2 3.2 0 4.4l-7.5 4.3-.1.1z"
-              fill="url(#svgcid--qj0fwh-s7c4e8)"
-            ></path>
-            <linearGradient
-              gradientTransform="matrix(1 0 0 -1 -4.357 -175.217)"
-              y2="-201.946"
-              x2="-90.988"
-              y1="-217.604"
-              x1="-106.646"
-              gradientUnits="userSpaceOnUse"
-              id="svgcid-98kv97-7egd8e"
-            >
-              <stop offset="0" stopColor="#32a071"></stop>
-              <stop offset=".069" stopColor="#2da771"></stop>
-              <stop offset=".476" stopColor="#15cf74"></stop>
-              <stop offset=".801" stopColor="#06e775"></stop>
-              <stop offset="1" stopColor="#00f076"></stop>
-            </linearGradient>
-            <path
-              d="M41.8-91.2 19.3-104c-1.3-.8-2.5-.7-3.2.1l19.2 19.2 6.5-6.5z"
-              fill="url(#svgcid-98kv97-7egd8e)"
-            ></path>
-            <path
-              d="M41.6-78.3 19.3-65.6c-1.3.7-2.4.7-3.1 0l-.1.1.1.1c.7.6 1.8.7 3.1 0l22.5-12.8-.2-.1z"
-              opacity=".2"
-            ></path>
-            <path
-              d="M16.1-65.7c-.4-.5-.7-1.2-.7-2.2v.2c0 1 .3 1.7.7 2.2l.1-.1-.1-.1z"
-              opacity=".122"
-            ></path>
-            <path
-              d="m49.3-82.7-7.7 4.4.1.1 7.5-4.3c1.1-.6 1.6-1.4 1.6-2.2.1.7-.5 1.5-1.5 2z"
-              opacity=".122"
-            ></path>
-            <path
-              d="m19.3-103.8 30 17.1c1 .6 1.5 1.3 1.6 2 0-.8-.5-1.6-1.6-2.2l-30-17.1c-2.2-1.2-3.9-.2-3.9 2.3v.2c0-2.5 1.7-3.5 3.9-2.3z"
-              opacity=".251"
-              fill="#ffffff"
-            ></path>
-            <path
-              d="M236.5 51.5v-41c0-.7.1-1.3.2-2 .1-.6.3-1.3.6-1.9.3-.6.6-1.2.9-1.7.4-.5.8-1.1 1.2-1.5.5-.5 1-.9 1.5-1.2.5-.4 1.1-.7 1.7-.9.6-.3 1.2-.4 1.9-.6.6-.1 1.3-.2 2-.2h187c.7 0 1.3.1 2 .2.6.1 1.3.3 1.9.6.6.3 1.2.6 1.7.9.5.4 1.1.8 1.5 1.2.5.5.9 1 1.2 1.5.4.5.7 1.1.9 1.7s.4 1.2.6 1.9c.1.6.2 1.3.2 2v41c0 .7-.1 1.3-.2 2-.1.6-.3 1.3-.6 1.9-.3.6-.6 1.2-.9 1.7s-.8 1.1-1.2 1.5c-.5.5-1 .9-1.5 1.2-.5.4-1.1.7-1.7.9-.6.3-1.2.4-1.9.6s-1.3.2-2 .2h-187c-.7 0-1.3-.1-2-.2-.6-.1-1.3-.3-1.9-.6-.6-.3-1.2-.6-1.7-.9-.5-.4-1.1-.8-1.5-1.2-.5-.5-.9-1-1.2-1.5s-.7-1.1-.9-1.7c-.3-.6-.4-1.2-.6-1.9-.1-.7-.2-1.3-.2-2z"
-              fill="#ffffff"
-            ></path>
-            <path
-              d="M236.5 52V10c0-.6.1-1.2.2-1.9.1-.6.3-1.2.5-1.8.2-.6.5-1.1.9-1.6.3-.5.7-1 1.2-1.4.4-.4.9-.8 1.4-1.2.5-.3 1.1-.6 1.6-.9.6-.2 1.2-.4 1.8-.5.6-.1 1.2-.2 1.9-.2h188c.6 0 1.2.1 1.9.2.6.1 1.2.3 1.8.5.6.2 1.1.5 1.6.9.5.3 1 .7 1.4 1.2.4.4.8.9 1.2 1.4s.6 1.1.9 1.6c.2.6.4 1.2.5 1.8.1.6.2 1.2.2 1.9v42c0 .6-.1 1.2-.2 1.9-.1.6-.3 1.2-.5 1.8-.2.6-.5 1.1-.9 1.6-.3.5-.7 1-1.2 1.4-.4.4-.9.8-1.4 1.2-.5.3-1.1.6-1.6.9-.6.2-1.2.4-1.8.5-.6.1-1.2.2-1.9.2H246c-.6 0-1.2-.1-1.9-.2-.6-.1-1.2-.3-1.8-.5s-1.1-.5-1.6-.9c-.5-.3-1-.7-1.4-1.2-.4-.4-.8-.9-1.2-1.4-.3-.5-.6-1.1-.9-1.6-.2-.6-.4-1.2-.5-1.8-.1-.7-.2-1.3-.2-1.9z"
-              fill="none"
-              stroke="#a5a5a5"
-            ></path>
-            <path d="M290.2 32.1c0-4.4 3.6-6.5 3.8-6.6-2.1-3-5.3-3.4-6.4-3.4-2.7-.3-5.3 1.6-6.7 1.6-1.4 0-3.5-1.6-5.8-1.5-2.9 0-5.7 1.7-7.2 4.4-3.1 5.4-.8 13.2 2.2 17.5 1.5 2.1 3.2 4.5 5.5 4.4 2.2-.1 3-1.4 5.7-1.4s3.4 1.4 5.7 1.4c2.4 0 3.9-2.1 5.3-4.3 1.7-2.4 2.4-4.8 2.4-4.9.1-.2-4.5-1.9-4.5-7.2z"></path>
-            <path d="M285.8 19.2c1.2-1.5 2-3.5 1.8-5.6-1.7.1-3.9 1.2-5.1 2.7-1.1 1.3-2.1 3.4-1.8 5.4 1.9.1 3.9-1.1 5.1-2.5z"></path>
-            <path d="M325.1 35.1h-6.2l-1.5 4.4h-2.6l5.9-16.3h2.7l5.9 16.3h-2.7l-1.5-4.4zm-5.6-2h4.9l-2.4-7.2h-.1l-2.4 7.2z"></path>
-            <path d="M342 33.6c0 3.7-2 6.1-5 6.1-1.7 0-3-.8-3.7-2.1h-.1v5.9h-2.4V27.6h2.4v2c.7-1.3 2.1-2.1 3.8-2.1 3 0 5 2.4 5 6.1zm-2.5 0c0-2.4-1.2-4-3.1-4-1.9 0-3.1 1.6-3.1 4s1.3 4 3.1 4 3.1-1.6 3.1-4z"></path>
-            <path d="M355.1 33.6c0 3.7-2 6.1-5 6.1-1.7 0-3-.8-3.7-2.1h-.1v5.9h-2.4V27.6h2.4v2c.7-1.3 2.1-2.1 3.8-2.1 3 0 5 2.4 5 6.1zm-2.5 0c0-2.4-1.2-4-3.1-4-1.9 0-3.1 1.6-3.1 4s1.3 4 3.1 4 3.1-1.6 3.1-4z"></path>
-            <path d="M363.8 35c.2 1.6 1.8 2.7 3.9 2.7 2.1 0 3.5-1.1 3.5-2.5 0-1.3-.9-2-3-2.5l-2.1-.5c-3-.7-4.4-2.1-4.4-4.4 0-2.8 2.5-4.7 5.9-4.7 3.4 0 5.8 1.9 5.9 4.7H371c-.1-1.6-1.5-2.6-3.5-2.6s-3.3 1-3.3 2.4c0 1.2.9 1.8 3 2.4l1.8.4c3.3.8 4.7 2.1 4.7 4.5 0 3.1-2.4 5-6.3 5-3.6 0-6.1-1.9-6.2-4.8h2.6z"></path>
-            <path d="M379.1 24.8v2.8h2.3v1.9h-2.3v6.6c0 1 .5 1.5 1.4 1.5.2 0 .6 0 .8-.1v1.9c-.3.1-.8.1-1.4.1-2.4 0-3.3-.9-3.3-3.2v-6.8h-1.7v-1.9h1.7v-2.8h2.5z"></path>
-            <path d="M382.7 33.6c0-3.7 2.2-6.1 5.6-6.1 3.4 0 5.6 2.4 5.6 6.1 0 3.8-2.2 6.1-5.6 6.1s-5.6-2.4-5.6-6.1zm8.8 0c0-2.6-1.2-4.1-3.2-4.1s-3.2 1.5-3.2 4.1c0 2.6 1.2 4.1 3.2 4.1s3.2-1.5 3.2-4.1z"></path>
-            <path d="M396 27.6h2.3v2h.1c.4-1.3 1.5-2.1 2.9-2.1.4 0 .6 0 .8.1v2.3c-.2-.1-.6-.1-1.1-.1-1.6 0-2.5 1.1-2.5 2.7v7.1H396v-12z"></path>
-            <path d="M413.3 36c-.3 2.2-2.4 3.6-5.1 3.6-3.5 0-5.6-2.3-5.6-6s2.2-6.2 5.5-6.2 5.4 2.3 5.4 5.9v.8h-8.4v.1c0 2 1.3 3.4 3.2 3.4 1.4 0 2.4-.6 2.7-1.7h2.3zm-8.2-3.5h5.9c-.1-1.8-1.2-3-2.9-3s-2.9 1.2-3 3z"></path>
-            <path d="M424.2-115.6H251c-1 0-2.1.1-3.1.3-1 .2-2 .5-2.9 1-.9.5-1.8 1.1-2.5 1.8s-1.3 1.6-1.8 2.5c-.5.9-.8 1.9-1 2.9-.2 1-.2 2-.3 3.1v38.4c0 1 .1 2.1.3 3.1.2 1 .5 2 1 2.9.5.9 1.1 1.8 1.8 2.5s1.6 1.3 2.5 1.8c.9.5 1.9.8 2.9 1 1 .2 2.1.2 3.1.3h176.3c1 0 2.1-.1 3.1-.3 1-.2 2-.5 2.9-1 .9-.5 1.8-1.1 2.5-1.8s1.3-1.6 1.8-2.5c.5-.9.8-1.9 1-2.9.2-1 .2-2 .3-3.1V-104c0-1-.1-2.1-.3-3.1-.2-1-.5-2-1-2.9-.5-.9-1.1-1.8-1.8-2.5s-1.6-1.3-2.5-1.8c-.9-.5-1.9-.8-2.9-1-1-.2-2-.2-3.1-.3h-3.1z"></path>
-            <path
-              d="M252.4-55.4H251c-.9 0-1.9-.1-2.9-.3-.9-.2-1.8-.4-2.5-.8-.8-.4-1.5-.9-2.1-1.6-.6-.6-1.2-1.3-1.6-2.1-.4-.8-.7-1.6-.8-2.5-.2-1-.2-2.1-.3-2.9v-38.3c0-.8.1-1.9.3-2.9.2-.9.4-1.8.8-2.6s.9-1.5 1.6-2.1c.6-.6 1.4-1.2 2.2-1.6.8-.4 1.6-.7 2.5-.8 1-.2 2.1-.2 2.9-.3h176.2c.8 0 1.8.1 2.9.2.9.2 1.8.4 2.6.8s1.5.9 2.1 1.6c.6.6 1.2 1.3 1.6 2.2.4.8.7 1.6.8 2.5.2 1 .2 2 .3 2.9v38.4c0 .9-.1 1.9-.3 2.8-.2.9-.4 1.8-.8 2.6s-.9 1.5-1.6 2.1c-.6.6-1.4 1.2-2.2 1.6-.8.4-1.6.7-2.6.8-1 .2-2 .2-2.9.3H252.4z"
-              fill="#ffffff"
-            ></path>
-            <path d="M277.9-85c0-4.4 3.6-6.5 3.8-6.6-2.1-3-5.3-3.4-6.4-3.4-2.7-.3-5.3 1.6-6.7 1.6-1.4 0-3.5-1.6-5.8-1.5-2.9 0-5.7 1.7-7.2 4.4-3.1 5.4-.8 13.2 2.2 17.5 1.5 2.1 3.2 4.5 5.5 4.4 2.2-.1 3-1.4 5.7-1.4s3.4 1.4 5.7 1.4c2.4 0 3.9-2.1 5.3-4.3 1.7-2.4 2.4-4.8 2.4-4.9.1-.2-4.5-1.9-4.5-7.2z"></path>
-            <path d="M273.5-97.9c1.2-1.5 2-3.5 1.8-5.6-1.7.1-3.9 1.2-5.1 2.7-1.1 1.3-2.1 3.4-1.8 5.4 1.9.1 3.9-1.1 5.1-2.5z"></path>
-            <path d="M297.8-76.7h-3.9v5.4c1.6 0 3.3-.1 4.8-.4l.1 1.2c-1.2.2-3.4.4-5.5.4h-.8v-7.7h5.3v1.1zm3.2 2.2h1.7v1.2H301v5.9h-1.4v-11.5h1.4v4.4z"></path>
-            <path d="M314.1-73.3v1.1h-4.4v2.2h-1.4v-2.2h-4.6v-1.1h10.4zm-9.2-3c0-1.3 1.5-2.3 4-2.3 2.4 0 4 1 4 2.3 0 1.4-1.6 2.3-4 2.3s-4-1-4-2.3zm1.5 7.4h6.5v1.1H305V-71h1.4v2.1zm0-7.4c0 .7 1.1 1.2 2.5 1.2 1.5 0 2.5-.5 2.5-1.2 0-.8-1-1.2-2.5-1.2-1.4 0-2.5.4-2.5 1.2z"></path>
-            <path d="M325.8-70v1.1h-10.4V-70h4.5v-1.9h-3.1v-3.8h6.3v-1.4h-6.3v-1.1h7.8v3.7h-6.3v1.5h6.6v1.1h-3.5v1.9h4.4z"></path>
-            <path d="M337.5-69h-10.4v-1.1h10.4v1.1zm-1.4-7.8h-6.3v3.2h6.4v1.1h-7.8V-78h7.7v1.2z"></path>
-            <path d="M348.9-75h-6.7v-1.1h6.7v1.1zm-3.3 5.4c-1.6 0-2.7-1-2.7-2.4s1.2-2.4 2.7-2.4c1.6 0 2.8 1 2.8 2.4s-1.2 2.4-2.8 2.4zm1.7-7.4h-3.5v-1.2h3.5v1.2zm-1.7 6.2c.8 0 1.4-.5 1.4-1.3s-.6-1.3-1.4-1.3c-.8 0-1.4.5-1.4 1.3s.6 1.3 1.4 1.3zm5.6-3.4h1.6v1.2h-1.6v5.6h-1.4v-11.5h1.4v4.7z"></path>
-            <path d="m354.8-69.5-.8-1.1c2.7-1.3 4.3-3.6 4.5-5.9h-3.9v-1.2h5.4c-.1 3.4-1.7 6.3-5.2 8.2zm8.6 2.1H362v-11.5h1.4v11.5z"></path>
-            <path d="M302.7-90.1h-6.2l-1.5 4.4h-2.6l5.9-16.3h2.7l5.9 16.3h-2.7l-1.5-4.4zm-5.5-2.1h4.9l-2.4-7.2h-.1l-2.4 7.2z"></path>
-            <path d="M319.7-91.7c0 3.7-2 6.1-5 6.1-1.7 0-3-.8-3.7-2.1h-.1v5.9h-2.4v-15.8h2.4v2c.7-1.3 2.1-2.1 3.8-2.1 3 0 5 2.4 5 6zm-2.5 0c0-2.4-1.2-4-3.1-4-1.9 0-3.1 1.6-3.1 4s1.3 4 3.1 4 3.1-1.5 3.1-4z"></path>
-            <path d="M332.8-91.7c0 3.7-2 6.1-5 6.1-1.7 0-3-.8-3.7-2.1h-.1v5.9h-2.4v-15.8h2.4v2c.7-1.3 2.1-2.1 3.8-2.1 3 0 5 2.4 5 6zm-2.5 0c0-2.4-1.2-4-3.1-4-1.9 0-3.1 1.6-3.1 4s1.3 4 3.1 4 3.1-1.5 3.1-4z"></path>
-            <path d="M341.5-90.3c.2 1.6 1.8 2.7 3.9 2.7 2.1 0 3.5-1.1 3.5-2.5 0-1.3-.9-2-3-2.5l-2.1-.5c-3-.7-4.4-2.1-4.4-4.4 0-2.8 2.5-4.7 5.9-4.7 3.4 0 5.8 1.9 5.9 4.7h-2.5c-.1-1.6-1.5-2.6-3.5-2.6s-3.3 1-3.3 2.4c0 1.2.9 1.8 3 2.4l1.8.4c3.3.8 4.7 2.1 4.7 4.5 0 3.1-2.4 5-6.3 5-3.6 0-6.1-1.9-6.2-4.8h2.6z"></path>
-            <path d="M356.8-100.4v2.8h2.3v1.9h-2.3v6.6c0 1 .5 1.5 1.4 1.5.2 0 .6 0 .8-.1v1.9c-.3.1-.8.1-1.4.1-2.4 0-3.3-.9-3.3-3.2v-6.8h-1.7v-1.9h1.7v-2.8h2.5z"></path>
-            <path d="M360.4-91.7c0-3.7 2.2-6.1 5.6-6.1 3.4 0 5.6 2.4 5.6 6.1 0 3.8-2.2 6.1-5.6 6.1s-5.6-2.3-5.6-6.1zm8.8 0c0-2.6-1.2-4.1-3.2-4.1s-3.2 1.5-3.2 4.1c0 2.6 1.2 4.1 3.2 4.1s3.2-1.5 3.2-4.1z"></path>
-            <path d="M373.7-97.6h2.3v2h.1c.4-1.3 1.5-2.1 2.9-2.1.4 0 .6 0 .8.1v2.3c-.2-.1-.6-.1-1.1-.1-1.6 0-2.5 1.1-2.5 2.7v7.1h-2.4v-12z"></path>
-            <path d="M391-89.2c-.3 2.2-2.4 3.6-5.1 3.6-3.5 0-5.6-2.3-5.6-6s2.2-6.2 5.5-6.2 5.4 2.3 5.4 5.9v.8h-8.4v.1c0 2 1.3 3.4 3.2 3.4 1.4 0 2.4-.6 2.7-1.7h2.3zm-8.2-3.6h5.9c-.1-1.8-1.2-3-2.9-3s-2.9 1.3-3 3z"></path>
-            <path d="M401-101.7h1.9v15.3H401v-7.7h-1.4c-.2 2.7-1.3 5.1-3.5 5.1-2.5 0-3.5-2.8-3.5-5.9 0-3 1.1-5.8 3.5-5.8 2.2 0 3.3 2.2 3.5 4.8h1.4v-5.8zm-3.2 6.7c0-2.2-.5-4-1.7-4-1.1 0-1.6 1.7-1.6 4s.5 3.9 1.6 3.9c1.2.1 1.7-1.6 1.7-3.9zm8.5-7.1v16.4h-1.9v-16.4h1.9z"></path>
-            <path d="M414.9-98.2c0 3.3 1.5 5.7 4 7.1l-1.2 1.5c-1.5-.9-3.4-2.7-3.9-4.4-.4 1.8-2.3 3.9-3.9 4.8l-1.3-1.5c2.7-1.6 4.1-4.2 4.1-7.5v-2.5h2.1v2.5zm7.6 12.5h-2v-9h-3.3v-1.7h3.3v-5.6h2v16.3z"></path>
-          </g>
-        </svg>
-      </Box>
-    </a>
-  );
-};
 
 export default GoogleApp;
