+++ src/component/home/CatchphraseStyle.ts
