+++ src/component/home/ExplainBuyStyle.ts
