+++ src/component/home/ExplainMarketStyle.ts
