+++ src/component/home/ExplainSellStyle.ts
