+++ src/component/home/SloganStyle.ts
