+++ src/pages/howto/index.tsx
@@ -1,14 +1,14 @@
-import React from "react";
-import Footer from "../../common/global/Footer";
-import Section from "../../component/howto/Section1";
+import React from 'react';
+import Footer from '../../common/global/Footer';
+import Section from '../../component/howto/Section1';
 
 const Howto = () => {
-  return (
-    <>
-      <Section />
-      <Footer />
-    </>
-  );
+	return (
+		<>
+			<Section />
+			<Footer />
+		</>
+	);
 };
 
 export default Howto;
